Added CustomSlider to playground, Messed around with React Sandbox to show examples of usage, Continued adding tests and examples in JavaScript Sandbox.
Since Last Commit: I've been spending a lot of time working on
websites instead of my JavaScript studies. I'm happy to say that
most of my work is almost done, so I can spend more time studying.
The unfortunate thing is that it is not just my studies that have
taken a toll, but also SimpleSim.

There's a new project underway, one that will be heavily utilizing
React and Laravel. My wishes were to actually have the backend
middleware be JavaScript based (Express), but one of the partners
is a PHP developer... A PHP developer and a JavaScript developer
walks into a bar.. Nah, nevermind. Anyways, we will be using Laravel!

The good thing is, the front end is going to act nearly (if not exactly)
the same as it would with an Express middleware, so there is no real
learning curve to have. Hopefully I can get a new Sandbox built over
the weekend to make a ToDo list with React and Express. I'm curious
towards how WebSockets are going to work.

Self-Note #1: Start pushing at the end of each project to avoid
the large commits and to allow for clearer commit messages.

Self-Note #2: Failed to work on simplesim for quite a few days
(over a week now), that needs to change once possible.

Self-Note #3: Also haven't studied JavaScript in 3 days, this needs
to change as well.