(Heavily) rewrote simplesim. The goal is to get closer to a module based system, since the engine will heavily rely on that type of thinking anyways. I also made a simple browser wrapper that will allow me to do testing in the browser instead of trying to do testing in the console.
You'll notice that I have gulp-babel and babel-loader installed. The
reasoning behind this is because the web wrapper uses strictly babel
while SimpleSim itself actually uses webpack. There was a restructuring
of code, and you'll notice that the old code that contained the currency
system, wallet, and player are gone and now there are a few file changes
as well. The CurrencyHandler has been turned into a model while the handler
itself (ssCurrencyHandler.js) can be found in the handlers file.

The reason for this change is to promote the idea of flexibility. All
models are meant to be base files, from that you can create your new handlers
as well as new currencies. In theory, instead of having all your currencies
in one file, you would instead create a currencies folder and list all your
currencies in there.

One daunting task that I'd like to try to make easier (for example) is listing
all the currencies you want. Through a module based system, as one might
imagine, explicitly requiring all currencies can be cumbersome. So, we need
to figure out a way to make it simpler to add in an active currency.

It may get to a point where you have SimpleSim (The Engine) and then a tool
to make your life easier. For now, the process can be creating a CurrencyList
and in that CurrencyList, doing the horrid task of requiring a Currency file
and then adding it to the exported object with a name attached to the currency:

// Server/CurrencyList.js
{
let USD = require('../currencies/usd.js')
let Fish = require('../currencies/fish.js')

module.exports = {
USD: USD,
Fish: Fish
}
}

And then in the server config:

// Server/config.js
{
let CurrencyList = require('./CurrencyList.js')

var activeCurrencies = {
USD: CurrencyList.USD,
Fish: CurrencyList.Fish
}
}

The automation of this could be done through a SimpleSimServerConfigTool..